Add softdeletable model as python util
